% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{check.models}
\alias{check.betas}
\alias{check.models}
\title{Check density of the responses and betas and functional process}
\usage{
check.models(models = 1:12, comp = TRUE, times = TRUE, M = 1000)

check.betas(models = 1:12, comp = TRUE)
}
\arguments{
\item{models}{TODO}

\item{comp}{TODO}

\item{times}{TODO}

\item{M}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# Check models
check.models(models = 1:12, comp = TRUE, times = TRUE)
check.models(models = 1:12, comp = FALSE, times = TRUE)

# Check betas
check.betas(models = 1:12, comp = TRUE)
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

