% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{check.scenarios}
\alias{check.betas}
\alias{check.scenarios}
\title{Check density of the responses and functional processes}
\usage{
check.scenarios(scenarios = 1:12, composite = TRUE, times = TRUE,
  M = 1000)

check.betas(scenarios = 1:12, composite = TRUE, times = TRUE,
  est.beta = TRUE)
}
\arguments{
\item{scenarios}{a vector giving the simulation scenarios to be checked.}

\item{composite}{flag to indicate the generation of data according to a functional linear model with non-null coefficient (\code{TRUE}) or with a null coefficient (\code{FALSE}).}

\item{times}{flag to indicate whether to show the time employed computing each model.}

\item{M}{number of samples employed to estimate the densities of the response.}

\item{est.beta}{flag to indicate whether to plot the functional coefficient estimate.}
}
\value{
Nothing. The functions are called for producing diagnostic plots.
}
\description{
\code{check.scenarios} produces plots displaying th densities of the response for different deviations. \code{check.betas} displays a sample of the functional process, the functional coefficient associated and its estimation based on 100 observations.
}
\examples{
# Check scenarios and deviations
check.scenarios(scenarios = 1:12, composite = TRUE, times = TRUE)

# Check betas
check.betas(scenarios = 1:12, composite = TRUE)
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

