% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{m.dev}
\alias{m.dev}
\title{Deviations from the composite hypothesis}
\usage{
m.dev(X.fdata, type, delta, beta0, comp = TRUE)
}
\arguments{
\item{X.fdata}{functional observations in the class \code{\link[fda.usc]{fdata}}.}

\item{type}{TODO}

\item{delta}{TODO}

\item{beta0}{TODO}

\item{comp}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
mod <- r.cfs.2003(n = 100)
for (i in 1:5) {
  for (delta in 0:3) {
    dev <- m.dev(X.fdata = mod$X.fdata, type = i, beta0 = mod$beta.fdata, 
                 delta = delta, comp = TRUE)
  }
}
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

