% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic model}
\usage{
quadratic(X.fdata, B)
}
\arguments{
\item{X.fdata}{TODO}

\item{B}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
t <- seq(0, 1, len = 101)
X.fdata <- r.bridge(n = 100, t = t)$X.fdata
quadratic(X.fdata = X.fdata, B = outer(t, t, function(x, y) sin(2 * pi * x) + 
                                                            cos(2 * pi * y)))
}

