% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r.bridge}
\alias{r.bridge}
\title{fdata.bridge}
\usage{
r.bridge(n = 100, t = seq(0, 1, len = 101), b = c(2, 4, 5)/sqrt(2))
}
\arguments{
\item{n}{TODO}

\item{t}{TODO}

\item{b}{TODO}
}
\value{
TODO
}
\description{
Brownian Motion: eigenvectors: sqrt(2)*sin((k-.5)*pi*t) --- Lambda: 1/((k-.5)^2*pi^2) Brownian
Bridge: eigenvectors: sqrt(2)*sin(k*pi*t) --- Lambda: 1/(k^2*pi^2)
# Make Bridgebrownian to have a certain R2
}
\examples{
plot(r.bridge(n = 100)$X.fdata)
}
\author{
Manuel Febrero-Bande (\email{manuel.febrero@usc.es}) and Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

