% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r.cfs.2003}
\alias{r.cfs.2003}
\title{cfs.2003}
\usage{
r.cfs.2003(n = 100, t = seq(0, 1, len = 201), b = c(2, 4, 5)/sqrt(2),
  type = "a")
}
\arguments{
\item{n}{TODO}

\item{t}{TODO}

\item{b}{TODO}

\item{type}{TODO}
}
\value{
TODO
}
\description{
The function fdata.cfs.2003.a() generates simulated data of the functional linear linear model with
 The function data.cfs.2003.b() generates simulated data of the functional linear linear model with
scalar response, Cardot, Ferraty and Sarda (2003).
The function data.cfs.2003.c() generates simulated data of the functional linear model with scalar
response, Cardot, Ferraty and Sarda (2003), but with the inclusion of different eigenfunctions.
}
\examples{
plot(r.cfs.2003(n = 100, type = "a")$X.fdata)
}
\author{
Manuel Febrero-Bande (\email{manuel.febrero@usc.es}) and Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}
\references{
# TODO
}

