% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r.gbm}
\alias{r.gbm}
\title{Geometric brownian motion}
\usage{
r.gbm(n = 100, t = seq(0, 1, len = 101), S0 = 1)
}
\arguments{
\item{n}{TODO}

\item{t}{TODO}

\item{S0}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
plot(r.gbm(n = 100))
}

