% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r.hh.2006}
\alias{r.hh.2006}
\title{hh.2006}
\usage{
r.hh.2006(n = 100, t = seq(0, 1, len = 201), imod = 1, ncb = 20)
}
\arguments{
\item{n}{TODO}

\item{t}{TODO}

\item{imod}{TODO}

\item{ncb}{TODO}
}
\value{
TODO
}
\description{
The function data.hh.2006 generates simulated data of the functional linear model with scalar
}
\examples{
plot(r.hh.2006(n = 100)$X.fdata)
}
\author{
Manuel Febrero-Bande (\email{manuel.febrero@usc.es}) and Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}
\references{
# TODO
}

