% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{r.mod}
\alias{r.mod}
\title{Functional regression models for simple and composite hypothesis}
\usage{
r.mod(n, model, delta, R2 = 0.95, comp = TRUE)
}
\arguments{
\item{n}{TODO}

\item{model}{TODO}

\item{delta}{TODO}

\item{R2}{TODO}

\item{comp}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
for (i in 1:12) {
  for (delta in 0:3) {
    samp <- r.mod(n = 10, model = i, delta = delta, R2 = 0.95, comp = TRUE)
  }
}
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

