% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp.flm.test.R
\name{rp.flm.test}
\alias{rp.flm.test}
\title{Goodness-of fit test for the functional linear model using random projections}
\usage{
rp.flm.test(X.fdata, Y, beta0.fdata = NULL, est.method = "pc",
  p.criterion = "SIC", pmax = 10, p = NULL, type.basis = "bspline",
  B = 5000, n.proj = 10, verbose = TRUE, projs = 0.9,
  same.rwild = FALSE, ...)
}
\arguments{
\item{X.fdata}{functional observations in the class \code{\link[fda.usc]{fdata}}.}

\item{Y}{scalar responses for the FLM. Must be a vector with the same number of elements as functions are in \code{X.fdata}.}

\item{beta0.fdata}{functional parameter for the simple null hypothesis, in the \code{\link[fda.usc]{fdata}} class. The \code{argvals} and \code{rangeval} arguments of \code{beta0.fdata} must be the same of \code{X.fdata}. If \code{beta0.fdata=NULL} (default), the function will test for the composite null hypothesis.}

\item{est.method}{estimation method for \eqn{\beta}{\beta}, only used in the composite case. There are three methods:
\itemize{
  \item{"pc"}{if \code{p} is given, then \eqn{\beta}{\beta} is estimated by \code{\link[fda.usc]{fregre.pc}}. Otherwise, \code{p} is chosen using \code{\link[fda.usc]{fregre.pc.cv}} and the \code{p.criterion} criterion.}
  \item{"pls"}{if \code{p} is given, \eqn{\beta}{\beta} is estimated by \code{\link[fda.usc]{fregre.pls}}. Otherwise, \code{p} is chosen using \code{\link[fda.usc]{fregre.pls.cv}} and the \code{p.criterion} criterion.}
  \item{"basis"}{if \code{p} is given, \eqn{\beta}{\beta} is estimated by \code{\link[fda.usc]{fregre.basis}}. Otherwise, \code{p} is chosen using \code{\link[fda.usc]{fregre.basis.cv}} and the \code{p.criterion} criterion. Both in \code{\link[fda.usc]{fregre.basis}} and \code{\link[fda.usc]{fregre.basis.cv}}, the same basis for \code{basis.x} and \code{basis.b} is considered.}
}}

\item{p.criterion}{for \code{est.method} equal to \code{"pc"} or \code{"pls"}, either \code{"SIC"}, \code{"SICc"} or one of the criterions described in \code{\link[fda.usc]{fregre.pc.cv}}. For \code{"basis"} a value for \code{type.CV} in \code{\link[fda.usc]{fregre.basis.cv}} such as \code{GCV.S}.}

\item{pmax}{maximum size of the basis expansion to consider in when using \code{p.criterion}.}

\item{p}{number of elements for the basis representation of \code{beta0.fdata} and \code{X.fdata} with the \code{est.method} (only composite hypothesis). If not supplied, it is estimated from the data.}

\item{type.basis}{type of basis if \code{est.method = "basis"}.}

\item{B}{number of bootstrap replicates to calibrate the distribution of the test statistic.}

\item{n.proj}{vector with the number of projections to consider.}

\item{verbose}{whether to show or not information about the testing progress.}

\item{projs}{a \code{\link[fda.usc]{fdata}} object containing the random directions employed to project \code{X.fdata}. If numeric, the convenient value for \code{ncomp} in \code{\link{rdir.pc}}.}

\item{same.rwild}{wether to employ the same wild bootstrap residuals for different projections or not.}

\item{...}{further arguments passed to \code{\link[fda]{create.basis}} (not \code{rangeval} that is taken as the \code{rangeval} of \code{X.fdata}).}
}
\value{
An object with class \code{"htest"} whose underlying structure is a list containing the following components:
\itemize{
  \item{p.values.fdr}{A matrix of size \code{c(n.proj, 2)}, containing in each row the FDR p-values of the CvM and KS tests up to that projection.}
  \item{proj.statistics}{A matrix of size \code{c(n.proj, 2)} with the value of the test statistic on each projection.}
  \item{boot.proj.statistics}{An array of size \code{c(n.proj, 2, B)} with the values of the bootstrap test statistics for each projection.}
  \item{proj.p.values}{A matrix of size \code{c(n.proj, 2)}}
  \item{method}{Information about the test performed and the kind of estimation performed.}
  \item{B}{Number of bootstrap replicates used.}
  \item{n.proj}{Number of projections considered.}
  \item{projs}{Random directions employed to project \code{X.fdata}.}
  \item{type.basis}{Type of basis for \code{est.method = "basis"}.}
  \item{beta.est}{Estimated functional parameter \eqn{\hat\beta}{\hat\beta} in the composite hypothesis. For the simple hypothesis, \code{beta0.fdata}.}
  \item{p}{Number of basis elements considered for estimation of \eqn{\beta}{\beta}.}
  \item{p.criterion}{Criterion employed for selecting \code{p}.}
  \item{data.name}{The character string "Y = <X, b> + e"}
}
}
\description{
Tests the composite null hypothesis of a Functional Linear Model with scalar response (FLM),
\deqn{H_0:\,Y=\langle X,\beta\rangle+\epsilon\quad\mathrm{vs}\quad H_1:\,Y\neq\langle X,\beta\rangle+\epsilon.}{H_0: Y=<X,\beta>+\epsilon vs H_1: Y!=<X,\beta>+\epsilon.}
If \eqn{\beta=\beta_0}{\beta=\beta0} is provided, then the simple hypothesis \eqn{H_0:\,Y=\langle X,\beta_0\rangle+\epsilon}{H_0: Y=<X,\beta0>+\epsilon} is tested. The way of testing the null hypothesis is via a norm (Cramer-von Mises or Kolmogorov-Smirnov) in the empirical process indexed by the projections.
}
\details{
No NA's are allowed neither in the functional covariate nor in the scalar response.
}
\examples{
# Simulated example

set.seed(345678)
t <- seq(0, 1, l = 101)
n <- 100
X <- rproc2fdata(n = n, t = t, sigma = "OU")
beta0 <- fdata(mdata = cos(2 * pi * t) - (t - 0.5)^2, argvals = t,
               rangeval = c(0,1))
Y <- inprod.fdata(X, beta0) + rnorm(n, sd = 0.1)

# Test all cases
rp.flm.test(X.fdata = X, Y = Y, est.method = "pc", B = 1000) #  p-value = 0.83, p-value = 0.75
rp.flm.test(X.fdata = X, Y = Y, est.method = "pls", B = 1000) #  p-value = 0.5000, p-value = 0.5889
rp.flm.test(X.fdata = X, Y = Y, est.method = "basis", 
            p.criterion = fda.usc::GCV.S, B = 1000) #  p-value = 0.2, p-value = <2e-16

rp.flm.test(X.fdata = X, Y = Y, est.method = "pc", p = 5, B = 1000)
rp.flm.test(X.fdata = X, Y = Y, est.method = "pls", p = 5, B = 1000)
rp.flm.test(X.fdata = X, Y = Y, est.method = "basis", p = 5, B = 1000)

rp.flm.test(X.fdata = X, Y = Y, beta0.fdata = beta0, B = 1000)

rp.flm.test(X.fdata = X, Y = Y, est.method = "pc", p = 1, B = 1000)
rp.flm.test(X.fdata = X, Y = Y, est.method = "pls", p = 1, B = 1000)
rp.flm.test(X.fdata = X, Y = Y, est.method = "basis", p = 5, B = 1000)
\dontrun{
# Composite hypothesis: do not reject FLM
rp.test <- rp.flm.test(X.fdata = X, Y = Y, est.method = "pc", B = 1000)
pcvm.test <- flm.test(X.fdata = X, Y = Y, est.method = "pc", B = 1000,
                      plot.it = FALSE)
rp.test
pcvm.test

# Estimation of beta
par(mfrow = c(1, 3))
plot(X, main = "X")
plot(beta0, main = "beta")
lines(rp.test$beta.est, col = 2)
lines(pcvm.test$beta.est, col = 3)
plot(density(Y), main = "Density of Y", xlab = "Y", ylab = "Density")
rug(Y)

# Simple hypothesis: do not reject beta = beta0
rp.flm.test(X.fdata = X, Y = Y, beta0.fdata = beta0, B = 1000)
flm.test(X.fdata = X, Y = Y, beta0.fdata = beta0, B = 1000, plot.it = FALSE)

# Simple hypothesis: reject beta = beta0^2
rp.flm.test(X.fdata = X, Y = Y, beta0.fdata = beta0^2, B = 1000)
flm.test(X.fdata = X, Y = Y, beta0.fdata = beta0^2, B = 1000, plot.it = FALSE)

# AEMET dataset

# Load data
data(aemet)
wind.speed <- apply(aemet$wind.speed$data, 1, mean)
temp <- aemet$temp

# Remove the 5\% of the curves with less depth (i.e. 4 curves)
par(mfrow = c(1, 1))
res.FM <- depth.FM(temp, draw = TRUE)
qu <- quantile(res.FM$dep, prob = 0.05)
l <- which(res.FM$dep <= qu)
lines(aemet$temp[l], col = 3)

# Data without outliers
wind.speed <- wind.speed[-l]
temp <- temp[-l]

# Exploratory analysis: do not reject the FLM
rp.aemet <- rp.flm.test(X.fdata = temp, Y = wind.speed, B = 1000,
                        est.method = "pc")
pcvm.aemet <- flm.test(X.fdata = temp, Y = wind.speed, B = 1000,
                       est.method = "pc", plot.it = FALSE)
rp.aemet
pcvm.aemet

# Estimated betas
plot(rp.aemet$beta.est, col = 2)
lines(pcvm.aemet$beta.est, col = 3)

# B = 5000 for more precision on the calibration: do not reject the FLM
rp.flm.test(X.fdata = temp, Y = wind.speed, B = 5000, est.method = "pc")
flm.test(X.fdata = temp, Y = wind.speed, B = 5000, est.method = "pc",
         plot.it = FALSE)

# Simple hypothesis: rejection of beta0 = 0
zero <- fdata(mdata = rep(0, length(temp$argvals)), argvals = temp$argvals,
              rangeval = temp$rangeval)
flm.test(X.fdata = temp, Y = wind.speed, beta0.fdata = zero, B = 1000,
         plot.it = FALSE)
rp.flm.test(X.fdata = temp, Y = wind.speed, beta0.fdata = zero, B = 1000)

# Tecator dataset

# Load data
data(tecator)
absorp <- tecator$absorp.fdata
ind <- 1:129 # or ind <- 1:215
x <- absorp[ind, ]
y <- tecator$y$Fat[ind]

# Exploratory analysis for composite hypothesis with automatic choose of p
rp.tecat <- rp.flm.test(X.fdata = x, Y = y, est.method = "pc", B = 5000)
pcvm.tecat <- flm.test(X.fdata = x, Y = y, est.method = "pc", B = 5000,
                       plot.it = FALSE)
rp.tecat
pcvm.tecat

# Distribution of the CvM and KS p-values
hist(rp.tecat$proj.p.values[, 1], lwd = 2, breaks = seq(0, 1, l = 20),
     freq = FALSE, main = "CvM and KS projected p-values")
rug(rp.tecat$proj.p.values)

# Distribution of the PCvM statistic
plot(density(pcvm.tecat$boot.statistics), lwd = 2, main = "PCvM distribution",
     xlab = "PCvM*", ylab = "Density under the null")
rug(pcvm.tecat$boot.statistics)
abline(v = pcvm.tecat$statistic, col = 2, lwd = 2)
legend("topright", legend = "PCvM observed", lwd = 2, col = 2)

# Simple hypothesis: fixed p
zero <- fdata(mdata = rep(0, length(x$argvals)), argvals = x$argvals,
              rangeval = x$rangeval)

# Fixed p
rp.flm.test(X.fdata = x, Y = y, beta0.fdata = zero, B = 1000, p = 11)
flm.test(X.fdata = x, Y = y, beta0.fdata = zero, B = 1000, p = 11)

# Simple hypothesis, automatic election of p
rp.flm.test(X.fdata = x, Y = y, beta0.fdata = zero, B = 1000)
flm.test(X.fdata = x, Y = y, beta0.fdata = zero, B = 1000)
}
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}) and Manuel Febrero-Bande (\email{manuel.febrero@usc.es}).
}
\references{
Cuesta-Albertos, J.A., Garcia-Portugues, E., Gonzalez-Manteiga, W. and Febrero-Bande, M. (2016). Goodness-of-fit tests for the functional linear model based on randomly projected empirical processes. arXiv XXXX:XXXX. \url{https://arxiv.org/abs/XXXX.XXXX}
Garcia-Portugues, E., Gonzalez-Manteiga, W. and Febrero-Bande, M. (2014). A goodness-of-fit test for the functional linear model with scalar response. Journal of Computational and Graphical Statistics, 23(3), 761--778. \url{http://dx.doi.org/10.1080/10618600.2013.812519}
}

