% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{test.models}
\alias{test.betas}
\alias{test.models}
\title{Check density of the responses and betas and functional process}
\usage{
test.models(models = 1:12, comp = TRUE, times = TRUE, M = 1000)

test.betas(models = 1:12, comp = TRUE)
}
\arguments{
\item{models}{TODO}

\item{comp}{TODO}

\item{times}{TODO}

\item{M}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# Test models
test.models(models = 1:12, comp = TRUE, times = TRUE)
test.models(models = 1:12, comp = FALSE, times = TRUE)

# Test betas
test.betas(models = 1:12, comp = TRUE)
}
\author{
Eduardo Garcia-Portugues (\email{edgarcia@est-econ.uc3m.es}).
}

