% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{check.scenarios}
\alias{check.scenarios}
\alias{check.betas}
\title{Check density of the responses and functional processes}
\usage{
check.scenarios(scenarios = 1:12, composite = TRUE, times = TRUE,
  R2 = 0.95, M = 1000, main = FALSE)

check.betas(scenarios = 1:12, n = 100, R2 = 0.95, times = TRUE,
  est.beta = TRUE, main = FALSE)
}
\arguments{
\item{scenarios}{a vector giving the simulation scenarios to be checked.}

\item{composite}{flag to indicate the generation of data according to a functional linear model with non-null coefficient (\code{TRUE}) or with a null coefficient (\code{FALSE}).}

\item{times}{flag to indicate whether to show the time employed computing each model.}

\item{R2}{proportion of variance of the response \eqn{Y}{Y} explained by the linear model when \code{delta = 0}. This is used to compute the variance of the error \eqn{\varepsilon}{\epsilon} of the regression model.}

\item{M}{number of samples employed to estimate the densities of the response.}

\item{main}{whether to show a caption indicating the scenario or not.}

\item{n}{the sample size.}

\item{est.beta}{flag to indicate whether to plot the functional coefficient estimate.}
}
\value{
Nothing. The functions are called for producing diagnostic plots.
}
\description{
\code{check.scenarios} produces plots displaying th densities of the response for different deviations. \code{check.betas} displays a sample of the functional process, the functional coefficient associated and its estimation based on 100 observations.
}
\examples{
\dontrun{
# Check scenarios and deviations
set.seed(3257641)
check.scenarios(scenarios = c(1, 7, 3, 5, 6, 4, 8, 9, 12), composite = TRUE)

# Check betas
set.seed(3257641)
check.betas(scenarios = c(1, 7, 3, 5, 6, 4, 8, 9, 12), n = 100)
}
}
\author{
Eduardo García-Portugués (\email{edgarcia@est-econ.uc3m.es}).
}
