% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{m.dev}
\alias{m.dev}
\title{Deviations from functional linearity}
\usage{
m.dev(X.fdata, type, delta, eta, composite = TRUE)
}
\arguments{
\item{X.fdata}{functional observations in the class \code{\link[fda.usc]{fdata}}.}

\item{type}{kind of deviation, an index from \code{1} to \code{5}.}

\item{delta}{an index from \code{0} to \code{3} denoting the degree of departure of the data from the null hypothesis of functional linearity, encoded with \code{0}.}

\item{eta}{functional parameter employed when \code{type = 4}.}

\item{composite}{flag to indicate the generation of data according to a functional linear model with non-null coefficient (\code{TRUE}) or with a null coefficient (\code{FALSE}).}
}
\value{
A vector of length \code{length(X.fdata)} containing \eqn{\delta \Delta(\mathcal{X})}{\delta \Delta(X)}.
}
\description{
Deviations from functional linearity considered in the simulation study of Cuesta-Albertos et al. (2017).
}
\details{
The description of the deviations is detailed in the supplementary material of Cuesta-Albertos et al. (2017).
}
\examples{
dev <- list()
k <- 1
mod <- r.cfs.2003(n = 100)
for (i in 1:5) {
  for (delta in 0:3) {
    dev[[k]] <- m.dev(X.fdata = mod$X.fdata, type = i, eta = mod$beta.fdata,
                      delta = delta, composite = TRUE)
    k <- k + 1
  }
}
}
\references{
Cuesta-Albertos, J.A., García-Portugués, E., Febrero-Bande, M. and González-Manteiga, W. (2017). Goodness-of-fit tests for the functional linear model based on randomly projected empirical processes. arXiv:1701.08363. \url{https://arxiv.org/abs/1701.08363}
}
\author{
Eduardo García-Portugués (\email{edgarcia@est-econ.uc3m.es}).
}
