% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{r.cfs.2003}
\alias{r.cfs.2003}
\alias{r.hh.2006}
\alias{r.bridge}
\title{Utilities for the simulation of functional linear models}
\usage{
r.cfs.2003(n, t = seq(0, 1, len = 201), b = c(2, 4, 5)/sqrt(2),
  type = c("a", "b")[1])

r.hh.2006(n, t = seq(0, 1, len = 201), imod = 1)

r.bridge(n, t = seq(0, 1, len = 201), b = c(2, 4, 5)/sqrt(2))
}
\arguments{
\item{n}{number of functions to sample.}

\item{t}{time locations for the functional data.}

\item{b}{coefficients of the functional coefficient in the theoretical basis of principal components of the Brownian motion.}

\item{type}{either example \code{"a"} or \code{"b"} from Cardot et al. (2003).}

\item{imod}{either \code{1}, \code{2} or \code{3} for denoting models (i), (ii) and (iii) in Hall and Hosseini-Nasab (2006).}
}
\value{
A list with the following elements:
\describe{
  \item{\code{X.fdata}}{the sample of functional data, an \code{\link[fda.usc]{fdata}} object of length \code{n}.}
  \item{\code{beta.fdata}}{the functional coefficient, an \code{\link[fda.usc]{fdata}} object.}
}
}
\description{
The functions \code{r.cfs.2003}, \code{r.hh.2006} and \code{r.bridge} sample the functional covariate and construct functional coefficients for their use in functional linear models:
\itemize{
  \item{\code{r.cfs.2003} implements examples (a) and (b) in Section 5 of Cardot et al. (2003).}
  \item{\code{r.hh.2006} gives models (i), (ii) and (iii) in Section 5 of Hall and Hosseini-Nasab (2006).}
  \item{\code{r.bridge} samples a Brownian motion and creates a functional coefficient constructed from the eigenfunctions \eqn{\sqrt(2) * \sin(k t \pi)}{\sqrt2 * sin(k*t*\pi)}.}
}
}
\examples{
# Cardot et al. (2003)
plot(r.cfs.2003(n = 100, type = "a")$X.fdata)
plot(r.cfs.2003(n = 100, type = "b")$X.fdata)

# Hall and Hosseini-Nasab (2006)
plot(r.hh.2006(n = 100, imod = 1)$X.fdata)
plot(r.hh.2006(n = 100, imod = 2)$X.fdata)
plot(r.hh.2006(n = 100, imod = 3)$X.fdata)

# Sample bridge
plot(r.bridge(n = 100)$X.fdata)
}
\references{
Cardot, H., Ferraty, F., Sarda, P. (2003) Spline estimators for the functional linear model. Statistica Sinica, 13(3), 571--592. \url{http://www3.stat.sinica.edu.tw/statistica/oldpdf/a13n31.pdf}

Hall, P. and Hosseini-Nasab, M. (2006) On properties of functional principal components analysis. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 68(1), 109--126. \url{http://dx.doi.org/10.1111/j.1467-9868.2005.00535.x}
}
\author{
Manuel Febrero-Bande (\email{manuel.febrero@usc.es}) and Eduardo García-Portugués (\email{edgarcia@est-econ.uc3m.es}).
}
