% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{r.mod}
\alias{r.mod}
\title{Functional regression models for simple and composite hypothesis}
\usage{
r.mod(n, scenario, delta = 0, t = seq(0, 1, l = 201), R2 = 0.95,
  composite = TRUE)
}
\arguments{
\item{n}{the sample size.}

\item{scenario}{an index from \code{1} to \code{9} denoting the simulation scenario.}

\item{delta}{an index from \code{0} to \code{3} denoting the degree of departure of the data from the null hypothesis of functional linearity, encoded with \code{0}.}

\item{t}{time locations for the functional data.}

\item{R2}{proportion of variance of the response \eqn{Y}{Y} explained by the linear model when \code{delta = 0}. This is used to compute the variance of the error \eqn{\varepsilon}{\epsilon} of the regression model.}

\item{composite}{flag to indicate the generation of data according to a functional linear model with non-null coefficient (\code{TRUE}) or with a null coefficient (\code{FALSE}).}
}
\value{
A list with the following elements:
\describe{
  \item{\code{X.fdata}}{the sample of functional data, an \code{\link[fda.usc]{fdata}} object of length \code{n}.}
  \item{\code{Y}}{the scalar responses, a vector of length \code{n}.}
  \item{\code{beta.fdata}}{the functional coefficient, an \code{\link[fda.usc]{fdata}} object.}
}
}
\description{
Sampling from the functional linear models considered in the simulation study of Cuesta-Albertos et al. (2017).
}
\details{
The samples are generated from the regression model
\deqn{Y = \langle \mathcal{X}, \beta\rangle + \delta \Delta(\mathcal{X})+\varepsilon,}{Y = <X, \beta> + \delta \Delta(X)+\varepsilon,}
where \eqn{\delta \Delta(\mathcal{X})}{\delta \Delta(X)} is computed by \code{\link{m.dev}}. The description of the scenarios is detailed in the supplementary material of Cuesta-Albertos et al. (2017).
}
\examples{
# Generate samples for all scenarios
samp <- list()
k <- 1
for (i in 1:9) {
  for (delta in 0:3) {
    samp[[k]] <- r.mod(n = 10, scenario = i, delta = delta, R2 = 0.95, 
                       composite = TRUE)
    k <- k + 1
  }
}
}
\references{
Cuesta-Albertos, J.A., García-Portugués, E., Febrero-Bande, M. and González-Manteiga, W. (2017). Goodness-of-fit tests for the functional linear model based on randomly projected empirical processes. arXiv:1701.08363. \url{https://arxiv.org/abs/1701.08363}
}
\author{
Eduardo García-Portugués (\email{edgarcia@est-econ.uc3m.es}).
}
